soundplayer: retype currentFrame to NSInteger, and verify four methods
SoundPlayer -setCurrentFrame: clamps its argument to [0, totalFrames],
calling totalFrames twice on the over-range path, which the reconstruction
already had. Its ivar encodes q, so the property, the backing ivar, the
getter, the setter parameter, and a local were all long long where the
width rule asks for NSInteger.

StoreButtonView's two colour setters store the retained colour into their
ivar and then send setNeedsDisplay, so both are custom setters rather than
synthesised ones. RBTutorialManager +getStatus: forwards to
RBUserSettingData getTutorialStatus:.

Candidate selection now filters against the VERIFIED keys as well as the
checklist column, after the previous round picked three addresses that were
already recorded.

+++ Project/SoundPlayer.h
@@ -37,7 +37,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @ghidraAddress 0x3557c (getter)
  * @ghidraAddress 0x3550c (setter)
  */
-@property(nonatomic, assign) long long currentFrame;
+@property(nonatomic, assign) NSInteger currentFrame;
 /**
  * @brief Whether the voice loops its asset. The setter is ignored while the voice is playing.
  * @ghidraAddress 0x355ac (getter)
